backend/requirements.txt
requests~=2.31.0
matplotlib~=3.8.4
networkx~=3.4.2
ortools~=9.11.4210
python-socketio~=5.11.4
aiohttp~=3.11.7